Fix use_ssl option when set via ENV variable
Summary
-------

When specifying `AHN_RINGFLUX_USE_SSL=false` the `false` value is
read in as the string `"false"`. When passed to the InfluxDB client,
`"false"` evaluates to a "truthy" value and causes the client to
attempt to use a SSL connection. This adds a value transform so that
when `AHN_RINGFLUX_USE_SSL=false` the InfluxDB client will NOT attempt
to use a SSL connection.